Hold property-animation clock, weight and fade lanes as JavaScript numbers
The property-animation records kept the group clock (current, from and
to time, speed ratio), the group weight, the clip's frame rate and
duration, the weight-fade jobs, the mixer's quaternion reference and the
manager's pre-update delta in float, where the pin holds JavaScript
numbers (property-animation.ts createPointerAnimationGroup's controller,
weighted-pointer-mixer.ts advancePropertyGroupTime and refX..refW,
animation-weight-fade.ts). The lowered bodies read them widened and each
store rounded, so a clock advanced in float steps. They are double now,
end to end: the records, the lowered advance, fade and accumulate bodies
bind them directly (AnimationFloatLane and the round-trip copies are
gone), the entry points take double, the compiler emits frame rates,
play ranges, speed ratios, weights, fade durations and property seeks
at double width, and a frame-keyed key time divides in double and rounds
once. The glTF group weight the same fades write is double too.

Kept float where the pin itself stores into a Float32Array: key times
and values (createSampler's `input`/`output`), the evaluated sample
(`_pointerScratch`), the blend buckets (`values: new F32(arity)`), and
the scalar writer, which stores one of those sample lanes.

Audit: RDN-16.

+++ native/include/bblite/runtime.hpp
@@ -3104,7 +3104,7 @@ struct AnimationGroupRecord {
     std::uint32_t asset = invalid_handle;
     std::size_t clip = 0;
     /** `AnimationGroup.weight`: what the weighted mixer contributes it at. */
-    float weight = 1.0f;
+    double weight = 1.0;
     std::weak_ptr<PropertyAnimationManagerRecord> animation_owner;
     double duration = 0;
     double frame_rate = 0;
@@ -6014,7 +6014,7 @@ void off_visibility_change(Engine& engine, std::size_t identity);
 #if !defined(BBLITE_HAS_ANIMATION) || BBLITE_HAS_ANIMATION
 #include <bblite/runtime/animation-api.hpp>
 #endif
-void set_animation_weight(Engine& engine, AnimationGroupHandle group, float weight);
+void set_animation_weight(Engine& engine, AnimationGroupHandle group, double weight);
 void go_to_frame(Engine& engine, AnimationGroupHandle group, float frame, bool with_engine);
 void play_animation(Engine& engine, AnimationGroupHandle group);
 void pause_animation(Engine& engine, AnimationGroupHandle group);
